test(harness): lock invariant registry fixtures
